Remove Rina's NavBar.js. Rename Standford's NavBar2.js to NavBar.js. Remove unused empty EditorPage.js.
