--- imports/ui/EditorPage.js
@@ -1,6 +1,0 @@
-import React, { Component } from 'react';
-class EditorPage extends Component {
-
-}
-
-export default EditorPage;
